day 23 isize -> i16

+++ solutions/day23/src/coordinate.rs
@@ -12,12 +12,12 @@ pub enum Direction {
 
 #[derive(Copy, Clone, Debug, Hash, Eq, PartialEq)]
 pub struct Coordinate {
-    pub x: isize,
-    pub y: isize,
+    pub x: i16,
+    pub y: i16,
 }
 
 impl Coordinate {
-    pub fn new(x: isize, y: isize) -> Self {
+    pub fn new(x: i16, y: i16) -> Self {
         Self { x, y }
     }
